move code to library
