+++ assets/accents/clown.ron
@@ -1,35 +1,38 @@
 (
     name: "clown",
-    patterns: [
-        ("$", Any([
-            Simple(" HONK!"),
-            Simple(" HONK HONK!"),
-        ])),
-    ],
-    severities: {
-        1: Extend((
-            patterns: [
-                ("$", Any([
-                    Simple(" HONK HONK!"),
-                    Simple(" HONK HONK HONK!"),
-                ])),
-            ],
-        )),
-        2: Extend((
-            patterns: [
-                ("$", Any([
-                    Simple(" HONK HONK HONK!"),
-                    Simple(" HONK HONK HONK HONK!"),
-                ])),
-            ],
-        )),
-        3: Extend((
-            patterns: [
-                ("$", Any([
-                    Simple(" HONK HONK HONK HONK!"),
-                    Simple(" HONK HONK HONK HONK HONK!!!!!!!"),
-                ])),
-            ],
-        )),
-    },
+    description: "Loud and honking",
+    rules: (
+        patterns: [
+            ("$", Any([
+                Simple(" HONK!"),
+                Simple(" HONK HONK!"),
+            ])),
+        ],
+        severities: {
+            1: Extend((
+                patterns: [
+                    ("$", Any([
+                        Simple(" HONK HONK!"),
+                        Simple(" HONK HONK HONK!"),
+                    ])),
+                ],
+            )),
+            2: Extend((
+                patterns: [
+                    ("$", Any([
+                        Simple(" HONK HONK HONK!"),
+                        Simple(" HONK HONK HONK HONK!"),
+                    ])),
+                ],
+            )),
+            3: Extend((
+                patterns: [
+                    ("$", Any([
+                        Simple(" HONK HONK HONK HONK!"),
+                        Simple(" HONK HONK HONK HONK HONK!!!!!!!"),
+                    ])),
+                ],
+            )),
+        },
+    ),
 )
