--- assets/accents/example.ron
@@ -1,82 +1,0 @@
-// regex syntax: https://docs.rs/regex/latest/regex/#syntax
-// regex flags: https://docs.rs/regex/latest/regex/#grouping-and-flags
-//
-// substitues pattern matches with replacements. all rules are applied from top to bottom
-// by default, m and i flags are added to every pattern which means ^ and $ match beginning
-// and end of line, regexses ignore case. you can opt out of these by doing "(?-mi)"
-// i is only added when pattern is fully composed of lowercase ascii letters
-(
-    // accent name, the only required field
-    name: "example",
-
-    // on by default, tries to match input case with output after each rule
-    // for example, if you replaced "HELLO" with "bye", it would use "BYE" instead
-    normalize_case: true,
-
-    // pairs of (regex, replacement)
-    // this is same as `patterns` except that each regex is surrounded with \b to avoid copypasting.
-    // `words` are applied before `patterns`
-    words: [
-        // this is the simplest rule to replace all "windows" words (separated by regex \b) 
-        // occurences with "linux", case sensitive
-        ("windows", Simple("linux")),
-        // this replaces word "OS" with one of replacements, with equal probability
-        ("os", Any([
-            Simple("Ubuntu"),
-            Simple("Arch"),
-            Simple("Gentoo"),
-        ])),
-    ],
-
-    // pairs of (regex, replacement)
-    // this is same as `words` except these are used as is, without \b
-    patterns: [
-        // inserts one of the honks. first value of `Weights` is relative weight. higher is better
-        ("$", Weights([
-            (32, Simple(" HONK!")),
-            (16, Simple(" HONK HONK!")),
-            (08, Simple(" HONK HONK HONK!")),
-            // ultra rare sigma honk - 1 / 56
-            (01, Simple(" HONK HONK HONK HONK!!!!!!!!!!!!!!!")),
-        ])),
-    ],
-
-    // accent can be used with severity (non negative value). higher severities can either extend
-    // lower level or completely replace it.
-    // default severity is 0. higher ones are defined here
-    severities: {
-        // extends previous severity (level 0, base one in this case), adding additional rules
-        // below existingones. words and patterns keep their relative order though - words are
-        // processed first
-        1: Extend(
-            (
-                words: [
-                    // even though we are extending, defining same rule will overwrite result.
-                    // relative order of rules remain the same: "windows" will remain first
-                    ("windows", Simple("windoos")),
-                ],
-
-                // extend patterns, adding 1 more rule
-                patterns: [
-                    // replacements can be nested arbitrarily
-                    ("[A-Z]", Weights([
-                        // 50% to replace capital letter with one of the Es
-                        (1, Any([
-                            Simple("E"),
-                            Simple("Ē"),
-                            Simple("Ê"),
-                            Simple("Ë"),
-                            Simple("È"),
-                            Simple("É"),
-                        ])),
-                        // 50% to do nothing, no replacement
-                        (1, Noop),
-                    ])),
-                ],
-            ),
-        ),
-
-        // replace severity 1 entirely. in this case with nothing. remove all rules on severity 2+
-        2: Replace(()),
-    },
-)
