+++ assets/accents/french.ron
@@ -1,90 +1,93 @@
 (
     name: "french",
-    words: [
-        ("a", Any([
-            Simple("un"),
-            Simple("une"),
-        ])),
-        ("am", Simple("suis")),
-        ("and", Simple("et")),
-        ("the", Any([
-            Simple("les"),
-            Simple("la"),
-            Simple("le"),
-        ])),
-        ("for", Simple("pour")),
-        ("of", Simple("de")),
-        ("my", Any([
-            Simple("mon"),
-            Simple("ma"),
-        ])),
-        ("very", Simple("très")),
-        ("want", Simple("vouloir")),
-        ("with", Simple("avec")),
-        ("i'm", Simple("je suis")),
-        ("i am", Simple("je suis")),
-        ("i", Simple("je")),
-        ("good", Simple("bon")),
-        ("bad", Simple("mal")),
-        ("spicy", Any([
-            Simple("épicé"),
-            Simple("épicée"),
-        ])),
-        ("yes", Simple("oui")),
-        ("no", Simple("non")),
-        ("why", Simple("pourquoi")),
-        ("what'?s", Simple("quel est")),
-        ("who'?s", Simple("qui est")),
-        ("hello", Any([
-            Simple("'allô"),
-            Simple("bonjour"),
-            Simple("salut"),
-        ])),
-        ("bye", Any([
-            Simple("bon voyage"),
-            Simple("adieu"),
-            Simple("au revoir"),
-        ])),
-        ("thanks", Simple("merci")),
-        ("assistant", Simple("ravageur")),
-        ("assistants", Simple("ravageurs")),
-        ("captain", Simple("capitaine")),
-        ("cook", Any([
-            Simple("cuisinier"),
-            Simple("cuisinière"),
-        ])),
-        ("enemy", Any([
-            Simple("silly english dog"),
-            Simple("ennemi"),
-            Simple("ennemie"),
-        ])),
-        ("friend", Simple("ami")),
-        ("friends", Simple("amis")),
-        ("greytider?", Simple("gitans")),
-        ("changeling", Simple("changeur")),
-        ("wizard", Simple("sorcier")),
-        ("(op|operative)", Simple("boche")),
-        ("(op|operative)s", Simple("boches")),
-        ("cheese", Any([
-            Simple("brie"),
-            Simple("roquefort"),
-            Simple("camembert"),
-        ])),
-        ("bread", Simple("baguette")),
-        ("tomato", Simple("tomate")),
-        ("wine", Simple("vin")),
-        ("traitor", Simple("traitre")),
-        ("maint", Simple("banlieues")),
-        ("nuke", Any([
-            Simple("grand bombe"),
-            Simple("la baguette ultime"),
-        ])),
-        ("shit", Simple("merde")),
-        ("urity", Simple("urite")),
-        ("security", Simple("securite")),
-        ("shitsec", Any([
-            Simple("gendarmerie"),
-            Simple("keufs"),
-        ])),
-    ],
+    description: "French",
+    rules: (
+        words: [
+            ("a", Any([
+                Simple("un"),
+                Simple("une"),
+            ])),
+            ("am", Simple("suis")),
+            ("and", Simple("et")),
+            ("the", Any([
+                Simple("les"),
+                Simple("la"),
+                Simple("le"),
+            ])),
+            ("for", Simple("pour")),
+            ("of", Simple("de")),
+            ("my", Any([
+                Simple("mon"),
+                Simple("ma"),
+            ])),
+            ("very", Simple("très")),
+            ("want", Simple("vouloir")),
+            ("with", Simple("avec")),
+            ("i'm", Simple("je suis")),
+            ("i am", Simple("je suis")),
+            ("i", Simple("je")),
+            ("good", Simple("bon")),
+            ("bad", Simple("mal")),
+            ("spicy", Any([
+                Simple("épicé"),
+                Simple("épicée"),
+            ])),
+            ("yes", Simple("oui")),
+            ("no", Simple("non")),
+            ("why", Simple("pourquoi")),
+            ("what'?s", Simple("quel est")),
+            ("who'?s", Simple("qui est")),
+            ("hello", Any([
+                Simple("'allô"),
+                Simple("bonjour"),
+                Simple("salut"),
+            ])),
+            ("bye", Any([
+                Simple("bon voyage"),
+                Simple("adieu"),
+                Simple("au revoir"),
+            ])),
+            ("thanks", Simple("merci")),
+            ("assistant", Simple("ravageur")),
+            ("assistants", Simple("ravageurs")),
+            ("captain", Simple("capitaine")),
+            ("cook", Any([
+                Simple("cuisinier"),
+                Simple("cuisinière"),
+            ])),
+            ("enemy", Any([
+                Simple("silly english dog"),
+                Simple("ennemi"),
+                Simple("ennemie"),
+            ])),
+            ("friend", Simple("ami")),
+            ("friends", Simple("amis")),
+            ("greytider?", Simple("gitans")),
+            ("changeling", Simple("changeur")),
+            ("wizard", Simple("sorcier")),
+            ("(op|operative)", Simple("boche")),
+            ("(op|operative)s", Simple("boches")),
+            ("cheese", Any([
+                Simple("brie"),
+                Simple("roquefort"),
+                Simple("camembert"),
+            ])),
+            ("bread", Simple("baguette")),
+            ("tomato", Simple("tomate")),
+            ("wine", Simple("vin")),
+            ("traitor", Simple("traitre")),
+            ("maint", Simple("banlieues")),
+            ("nuke", Any([
+                Simple("grand bombe"),
+                Simple("la baguette ultime"),
+            ])),
+            ("shit", Simple("merde")),
+            ("urity", Simple("urite")),
+            ("security", Simple("securite")),
+            ("shitsec", Any([
+                Simple("gendarmerie"),
+                Simple("keufs"),
+            ])),
+        ],
+    ),
 )
