+++ crates/accents/Cargo.toml
@@ -6,8 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rand = "0.8.5"
-regex = "1.10.0"
+serde = "1.0.189"
 ron = "0.8.1"
-serde = { version = "1.0.189", features = ["derive"] }
-log = "0.4.20"
+pink_accents = { git = "https://github.com/Fogapod/pink_accents", rev = "fee5774c8dba0e600d5120e2030b405ccb89b184" }
